requirements.txt
APScheduler==3.6.3
asgiref==3.5.1
astroid==2.11.7
atomicwrites==1.4.0
attrs==21.4.0
backports.zoneinfo==0.2.1
Brotli==1.0.9
cachetools==4.2.2
certifi==2021.10.8
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.0.12
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
cssselect2==0.7.0
defusedxml==0.7.1
dill==0.3.5.1
Django==3.2.14
django-colorfield==0.7.2
django-cors-headers==3.12.0
django-debug-toolbar==2.2
django-filter==21.1
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-base64==2.0
drf-yasg==1.20.0
Faker==12.0.1
flake8==3.9.2
flake8-broken-line==0.4.0
flake8-docstrings==1.6.0
flake8-isort==4.1.1
flake8-plugin-utils==1.3.2
flake8-polyfill==1.0.2
flake8-return==1.1.3
fonttools==4.37.4
gobject==0.1.0
html5lib==1.1
idna==2.8
importlib-metadata==1.7.0
inflection==0.5.1
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mixer==7.1.2
more-itertools==8.12.0
oauthlib==3.2.1
packaging==21.3
pep8-naming==0.12.1
Pillow==9.2.0
platformdirs==2.5.2
pluggy==0.13.1
psycopg2-binary==2.8.6
py==1.11.0
pyaes==1.6.1
pycodestyle==2.7.0
pycparser==2.21
pydocstyle==6.1.1
pydyf==0.5.0
pyflakes==2.3.1
PyJWT==2.1.0
pylint==2.13.9
pyparsing==3.0.7
pyphen==0.13.0
Pyrogram==2.0.4
PySocks==1.7.1
pytest==6.2.4
pytest-django==4.4.0
pytest-pythonpath==0.7.3
python-dateutil==2.8.2
python-dotenv==0.19.0
python-telegram-bot==13.7
python3-openid==3.2.0
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
reportlab==3.6.11
requests==2.26.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.14.0
snowballstemmer==2.2.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sorl-thumbnail==12.7.0
sqlparse==0.4.2
telegram==0.0.1
testfixtures==6.18.5
tinycss2==1.1.1
toml==0.10.2
tomli==2.0.1
tornado==6.1
typed-ast==1.5.4
typing_extensions==4.1.1
tzdata==2022.1
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.25.11
wcwidth==0.2.5
webencodings==0.5.1
wrapt==1.14.1
zipp==3.7.0
zopfli==0.2.1
